Add some timezone calculations.
One very useful thing that Toptal doesn't natively provide is a way
to filter out those jobs which require more overlap than you're willing
to provide. It could do this, but the calculations are a bit involved;
timezones are complex and weird creatures.

Still, with the right libraries, they can be tamed, so we do that here.
It's not perfect--the numbers will shift if you get caught in between
different countries' implementations of daylight savings time--but even
then, it gives answers that are correct in the moment, just not in general.